feat: add evaluation.md (#188)
* move qianfan.pydantic package into qianfan.utils

* add evaluation.md

* add menu
