+++ src/qianfan/resources/console/utils.py
@@ -16,12 +16,12 @@
 Utils for console api
 """
 import functools
-from typing import Any, Awaitable, Callable
+from typing import Any, Awaitable, Callable, Tuple
 
 from qianfan import get_config
+from qianfan.errors import InvalidArgumentError
 from qianfan.resources.requestor.console_requestor import ConsoleAPIRequestor
 from qianfan.resources.typing import ParamSpec, QfRequest, QfResponse, RetryConfig
-from qianfan.utils import _get_console_ak_sk
 
 P = ParamSpec("P")
 
@@ -98,3 +98,21 @@ def async_console_api_request(
         )
 
     return inner
+
+
+def _get_console_ak_sk(pop: bool = True, **kwargs: Any) -> Tuple[str, str]:
+    """
+    extract ak and sk from kwargs
+    if not found in kwargs, will return value from global config and env variable
+    if `pop` is True, remove ak and sk from kwargs
+    """
+    ak = kwargs.get("ak", None) or get_config().ACCESS_KEY
+    sk = kwargs.get("sk", None) or get_config().SECRET_KEY
+    if ak is None or sk is None:
+        raise InvalidArgumentError("ak and sk cannot be empty")
+    if pop:
+        # remove ak and sk from kwargs
+        for key in ("ak", "sk"):
+            if key in kwargs:
+                del kwargs[key]
+    return ak, sk
