+++ src/qianfan/utils/__init__.py
@@ -25,9 +25,7 @@ from qianfan.utils.logging import (
 )
 from qianfan.utils.utils import (
     AsyncLock,
-    _get_console_ak_sk,
     _get_from_env_or_default,
-    _get_qianfan_ak_sk,
     _get_value_from_dict_or_var_or_env,
     _none_if_empty,
     _set_val_if_key_exists,
@@ -42,8 +40,6 @@ __all__ = [
     "log_warn",
     "enable_log",
     "disable_log",
-    "_get_console_ak_sk",
-    "_get_qianfan_ak_sk",
     "_get_from_env_or_default",
     "_get_value_from_dict_or_var_or_env",
     "_none_if_empty",
