+++ src/qianfan/utils/utils.py
@@ -22,9 +22,8 @@ import threading
 import uuid as uuid_lib
 from threading import current_thread
 from types import TracebackType
-from typing import Any, Dict, Optional, Tuple, Type
+from typing import Dict, Optional, Type
 
-from qianfan import get_config
 from qianfan.errors import InvalidArgumentError
 from qianfan.utils import log_info
 
@@ -109,42 +108,6 @@ def _none_if_empty(val: str) -> Optional[str]:
     if val == "" or val.lower() == "none":
         return None
     return val
-
-
-def _get_console_ak_sk(pop: bool = True, **kwargs: Any) -> Tuple[str, str]:
-    """
-    extract ak and sk from kwargs
-    if not found in kwargs, will return value from global config and env variable
-    if `pop` is True, remove ak and sk from kwargs
-    """
-    ak = kwargs.get("ak", None) or get_config().ACCESS_KEY
-    sk = kwargs.get("sk", None) or get_config().SECRET_KEY
-    if ak is None or sk is None:
-        raise InvalidArgumentError("ak and sk cannot be empty")
-    if pop:
-        # remove ak and sk from kwargs
-        for key in ("ak", "sk"):
-            if key in kwargs:
-                del kwargs[key]
-    return ak, sk
-
-
-def _get_qianfan_ak_sk(pop: bool = True, **kwargs: Any) -> Tuple[str, str]:
-    """
-    extract ak and sk from kwargs
-    if not found in kwargs, will return value from global config and env variable
-    if `pop` is True, remove ak and sk from kwargs
-    """
-    ak = kwargs.get("ak", None) or get_config().AK
-    sk = kwargs.get("sk", None) or get_config().SK
-    if ak is None or sk is None:
-        raise InvalidArgumentError("ak and sk cannot be empty")
-    if pop:
-        # remove ak and sk from kwargs
-        for key in ("ak", "sk"):
-            if key in kwargs:
-                del kwargs[key]
-    return ak, sk
 
 
 class AsyncLock:
